+++ bus_user/bus_serial.py
@@ -78,7 +78,7 @@ class TypeWrReturn(Enum):
 
 
 # =====================================================================================================================
-class BusSerial:
+class BusSerial_Base:
     ADDRESS: str = None
     TIMEOUT_READ: float = 0.2
     TIMEOUT_WRITE: float = 0.5
@@ -280,7 +280,7 @@ class BusSerial:
 
         result: List[str] = []
         for name in attempt_list:
-            if BusSerial(address=name).address_check_exists():
+            if BusSerial_Base(address=name).address_check_exists():
                 result.append(name)
 
         if _lock_port:
@@ -471,7 +471,9 @@ class BusSerial:
         history.print_io()
         return history
 
-    # CMD MAP =========================================================================================================
+
+# =====================================================================================================================
+class BusSerialWGetattr_Base(BusSerial_Base):
     def __getattr__(self, item) -> Callable[..., HistoryIO]:
         """if no exists attr/meth
 
@@ -480,7 +482,7 @@ class BusSerial:
 
         1. SHOW (optional) COMMANDS EXPLICITLY by annotations without values!
         -----------------------------------------------------------------
-            class MySerialDevice(BusSerial):
+            class MySerialDevice(BusSerial_Base):
                 IDN: Callable[[Any], TYPE__RW_ANSWER]
                 ADDR: Callable[[Any], TYPE__RW_ANSWER]
                 DUMP: Callable[[Any], TYPE__RW_ANSWER]
@@ -500,8 +502,8 @@ class BusSerial:
 # =====================================================================================================================
 if __name__ == "__main__":
     # see/use tests
-    # ports = BusSerial.detect_available_ports()
-    # obj = BusSerial(address=ports[0])
+    # ports = BusSerial_Base.detect_available_ports()
+    # obj = BusSerial_Base(address=ports[0])
     # obj.connect()
     pass